refactor(Banner): use Link for logo navigation
Update the Banner component so that clicking the logo uses Next.js <Link> instead of an inline onClick with window.location.href="/".

This was done to enable preservation of React states when the user navigates back and forth between pages as the previous implementation did full page reloads when navigating from the logo.